UCR Dota 2
These are the primary analysis MATLAB scripts. There are a few comments that describe the different components as they go along.
UCR_dota2.m determines the different vectors for the winrate and pickrate statsitcs, as well as their biases.

"               "_game_simulator.m is a simplistic loop of games where, each character is randomly selected for the two teams and the team with the higher winrate is considered to be the victor. From there, it logs that win, and begins to formulate a signal, with a little added noise for random variation, so that in the final script we have a series of game wins for two separate teams.

"               "_signal_generation.m is the final script, in which the team winrate and character winrate are combined, and each peak represents (in theory) the highest likelihood of a team win, with that particular character. This creates a vector for the Scrimp++ file, for analysis.